altered delete to destroy, added if else to destroy method.

+++ app/controllers/api/v1/books_controller.rb
@@ -37,10 +37,13 @@ class Api::V1::BooksController < ApplicationController
     end
   end
 
-  def delete
+  def destroy
     book = Book.find_by(id: params[:id])
-    book.destroy
-    # render json: book
+    unless book.nil?
+      book.destroy
+      render json: book
+    else
+      render json: { error: "Book not Found!" }, status: :unprocessable_entity
   end
 
   private
